[CFE] Expression compilation overwrites types to static types if from "dart:"
When the VM has an `int` it often represents it as a `_Smi` (which
extends `_IntegerImplementation` which implements `int`), but the
interface member for e.g. `+` on `_Smi` is `_IntegerImplementation.+`
and `_IntegerImplementation` is not special-cased in
`isSpecialCasedBinaryOperator` (but `int` and `num` is, so using `int`
having the interface member `num.+` makes `isSpecialCasedBinaryOperator`
return true if the type is `int` but false if the type is `_Smi`).

Not knowing which specialized types the VM could send for more general
types that are then specialized in the compiler this CL fixes the issue
(linked below) by using the static types of definitions from `dart:`.

Fixes https://github.com/dart-lang/sdk/issues/61894

Change-Id: I66cb2d44d5762719b14b51cb139fa0df6b00468f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459420
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>