analyzer_testing: Simplify caching mock files
Before this change, `_cacheFiles` was a function that took a Map as a
parameter, which was always empty, and returned `void`, because the
Map is an output parameter. This was kind of like a return type, but
with extra steps.

In this CL, I make the function return a Map, and then
`MockPackagesMixin._cachedFiles` can be made non-nullable, and
`late final`, with a simpler lifecycle.

Change-Id: Iae236e43de32d4ed82466cfa5661d266c52b14ae
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/460202
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>