[vm] Rework handling of section indices in the MachOWriter.
Store the index of the section in the load commands in the section.

Initialize it at the same time as the symbol tables are initialized,
since they are the primary client of this index.

Instead of storing the section index separately in symbols, which
means they must be updated after any reordering, just keep a pointer
to the owning section (or nullptr, if this is a global symbol).

This also allows us to lazily determine the final value written for
section-owned symbols using the owning section's memory address,
so now there is no finalization step needed for symbol tables.

TEST=ci

Cq-Include-Trybots: luci.dart.try:vm-aot-dwarf-linux-product-x64-try,vm-aot-linux-debug-x64-try,vm-aot-mac-debug-arm64-try
Change-Id: I87473a8fedf085e08492db62f133882c886bd9cb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/460045
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Tess Strickland <sstrickl@google.com>