[private named parameters] Implement in analyzer
I considered a few different approaches for how to model a formal parameter having a public and private name. Here's a simple example:

```dart
class C {
  int? _f;
  C({this._f});
}

main() {
  C(f: 1);
}
```

Here, it seems like the right approach is to give the formal parameter the public name since that's what you pass for the parameter. Then have resolution understand that the parameter can still map to a private instance variable.

But then consider:

```dart
class C {
  int? _f;
  int? x;
  C({this._f}) : x = _f;
}
```

The formal parameter for _f can also be accessed inside the initializer list and here it must have the private name.

So, really, the name of the formal parameter *is* the private name. It's what the user wrote. It's the name that matches the instance field. It's the name that can be used to access the formal parameter in the initializer list.

The *only* places the public name come into play are:

1. The name you use to pass an argument to that parameter in a named expression.

2. The generated Dartdoc for that constructor.

Given that, I though a simple solution might be to give FormalParameterElement another property for "argument name". In places where we are binding an argument to the parameter, we use that to figure out which parameter is being referred to. Later when I work on Dartdoc support, I'll use that when generating the constructor signature.

Everything else just keeps using the name.

Thoughts?

Change-Id: Id6ccf8d850480f141dc0dc6e26f7b185b60f6c3d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459663
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>
Auto-Submit: Bob Nystrom <rnystrom@google.com>