[messages] Extract computation of constantStyle.
Extracts the computation of `concreteClassName` and `staticType` from
`MessageWithAnalyzerCode.toAnalyzerCode` to a late field,
`constantStyle`, that encapsulates these values. The late field also
captures information about what kind of diagnostic constant should be
code generated, and if it should be a diagnostic constant that
supports `withArguments`, what the parameters of `withArguments`
should be.

This paves the way for a follow-up CL that will move the diagnostic
constants to top level declarations; this logic will require multiple
methods in `MessageWithAnalyzerCode` to refer to the constant style.

Change-Id: I6a6a69644ff69061186f61a6ba146916fbf4ef1c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461141
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>