Documenting our design choices
There are a couple of questions about the design of the analysis
server's features that have come up over the years. I think we want to
capture our decisions and the reasons for them so that we can refer
back to a doc rather than retyping the whole description over and over.
This is an attempt to capture a couple of those topics as a first step.

I'm not completely happy with the structure of the doc. I had originally
thought that I'd have a 'principles' directory to discuss the design
principles we follow, and a 'features' directory to discuss how those
principles are applied to the design of each feature. This initial doc
has both in one file because it wasn't clear how best to split it up.
It's hard to talk about principles in the absence of concrete examples.
I'll probably re-organize the text later, and am hoping you have some
ideas.

By writing this down I'm not trying to stop the discussion of how the
features should be designed. I'm just trying to capture the current
state of the design. Design is, and should be, fluid, and I'm happy to
update these docs as our design changes. So, please review this more
from a perspective of whether the current state is correctly represented
and less from a perspective of whether the current state is what it
ought to be.

Change-Id: I8e3e1de2ca01f3e18fdacc89317ddb0e5a98484e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/460904
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>