[messages] Add logic to generate analyzer diagnostics as top level constants.
Modifies the analyzer diagnostic message code generation logic so that
it's capable of generating diagnostic constants either:

- As static constants inside `DiagnosticCode`-derived classes such as
  `CompileTimeErrorCode`

- Or as top level constants inside of new files called
  `diagnostic.g.dart`.

The behavior is controlled through the constant
`generateTopLevelConstants`. This constant is currently `false`, which
means that diagnostic constants are still generated as static
constants inside `DiagnosticCode`-derived classes (as they have been
for quite some time). This CL introduces the new `diagnostic.g.dart`
files, but they are currently empty.

In a follow-up CL I will flip the flag to `true` and re-run the code
generation logic, and in a later CL, I'll remove the flag. The reason
for spreading the work across multiple CLs is twofold:

- To make code review easier

- To make merge conflicts easier to deal with (since the flag flip CL
  can easily be regenerated if merge conflicts arise).

Note that when the flag is set to `true`, static constants will still
be generated to the `DiagnosticCode`-derived classes to avoid breaking
code in the analyzer and related packages, but these constants will
simply redirect to the constants in the new `diagnostic.g.dart` files.

Change-Id: I6a6a696452d1a81a45010b374e898f0fa5fe5c9a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461260
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>