[messages] Write a tool to convert error references to `diag.`.
Adds the file `switch_to_toplevel_diagnostics.dart`. This is a
temporary utility that modifies references to diagnostic constants in
the analyzer and related packages, so that instead of referring to
constants in classes like `CompileTimeErrorCode`, etc., they refer to
top level constants in `diagnostic.dart` files using the import prefix
`diag`.

In a follow-up CL I will run this utility (which will actually make
the changes), and then in a follow-up CL after that, I will delete
it. The reason for splitting this up into multiple CLs is twofold:

- To allow easier code review.

- To make it easier to resolve merge conflicts in the CL that actually
  makes the changes, since it can easily be regenerated from scratch
  by re-running the tool.

Change-Id: I6a6a6964a6bfdf43227bfe02ba42a673901cacbc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461582
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>