[parser] Compute isConstructor before beginMethod
This moves the computation of which a method-like declaration is a constructor to before the listener call to beginMethod. This paves the way for creating a beginXConstructor call that fully separates listener handling of constructors and methods.

To avoid changes in the parsed AST, the computation deliberately matches the existing result, even though this is intertwined with error recovery.

Change-Id: I34ba68f9690e842a6c8baa26d9ba4b651230ad76
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461040
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>