[stable][analyzer] Dot shorthands: Error when calling an instance method using a shorthand.
Issue description: When attempting to use call an instance method `foo`
with dot shorthand syntax, no analyzer error would appear. The CFE
correctly produces an error, but we're missing the check on the
analyzer.

What is the fix: If we resolve to a method, do a quick check to make
sure it's static and not an instance method.

Why cherry-pick: Users should get this warning at compile time and not
have to run the program to see that their dot shorthand code is wrong.

Risk: Low, this fix has landed on the main channel and tests confirm
the behaviour.

Issue link(s): https://github.com/dart-lang/sdk/issues/61954

Bug: https://github.com/dart-lang/sdk/issues/61954
Change-Id: I0537adb0a0a02d982686af99a09ba91d979c09b9
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/461900
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462080
Reviewed-by: Alexander Thomas <athom@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>