[messages] Add package getter to MessageWithAnalyzerCode.
Adds the abstract getter `MessageWithAnalyzerCode.package` (and
implementations thereof). This returns the package into which this
error code will be generated.

Also adds a check to the `DiagnosticTables._` constructor to verify
that the values returned by this getter are the same as the values
returned by
`MessageWithAnalyzerCode.diagnosticClassInfo.file.package`, and
changes the code generation logic to use the former instead of the
latter where possible.

This paves the way for a follow-up CL that will remove
`MessageWithAnalyzerCode.diagnosticClassInfo`, as part of a larger
effort to remove the generated `DiagnosticCode`-derived classes
entirely.

Change-Id: I6a6a6964b03e53867f7dc7f7307e8e218c080b6c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462161
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>