Add access to the session logger on the insights pages
This adds a page on which users can start and stop the collection of
log entries and then copy and paste those entries either to send to us
or to write a scenario.

I changed my mind about the interaction style I wanted on this new page,
so I had to modify the in-memory session log sink slightly.

I also changed one of the labels to reflect the rename to the insights
pages.

Change-Id: I460668047d777f2ee98764ecff7845ea2d2fda99
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462244
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>