add analysis server scenarios, project generators, and a simple runner script
Currently supports a "git worktree" mode for running on local directories, as well as a "git clone" mode for cloning remote directories.

You can execute the script, but it will fail when it gets to replaying the logs due to the empty Json file, but this is a good start.

I think the next step here is to have some way of normalizing the logs so they are portable across machines/directories.

Change-Id: I8de5defe81c76f18c3bf3220e2a517a4247d4641
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462260
Commit-Queue: Jake Macdonald <jakemac@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>