[messages] Fix error_fix_status.yaml.
Adds missing `status` entries to the analysis server's
`error_fix_status.yaml` file.

The reason these were missing was because currently, the only
validation logic that checks that `status` entries are present is in
`pkg/linter/tool/machine.dart` (the `readFixStatusMap` function), and
it ignores entries that don't start with the class name prefix
`LintCode.`.

This paves the way for a follow-up CL that will remove the class name
prefixes from `error_fix_status.yaml` entirely (as part of a larger
effort to remove those classes from the codebase).

Note that the validation logic in `readFixStatusMap` is not really
very robust; it only checks that the `status` entries are strings. In
a follow-up CL I will add a more thorough check.

Change-Id: I6a6a69644f7cc07162f5614c87149d6e72b99d23
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462700
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>