Adjust the lexical lookup
This CL adjusts the lexical lookup such that an instance member
declaration in the lexically enclosing syntax is reported using null for
the getter and for the setter result. This ensures that the search ends
(we shouldn't search the library scope in this situation), and it
ensures that the given member is handled in the same way as if the
search had exhausted all levels and found nothing, which will then cause
further processing to prepend `this.` and resolve the identifier as an
instance member.

This change aligns the implemented behavior to match the language
specification more precisely, and it is necessary in order to get the
desired and specified behavior with lookups that originate in an
anonymous method.

Change-Id: Ic11c412b91827d57455a1e7febbb031b5aadad70
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459740
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Erik Ernst <eernst@google.com>