[messages] Better validation for error_fix_status.
Adds logic to `verify_error_fix_status.dart` to validate that every
entry is a YAML map and contains a valid status code.

The error reporting logic is rewritten to make use of the
`LocatedError` class so that if validation fails, it's not necessary
to describe where to find the erroneous entry; instead, the offending
path, line number, and column is printed as part of the exception
message.

Change-Id: I6a6a6964c8cc3094f81ad2b28dd444bfcb6c15c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463084
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>