[parser] Handle qualified names in new constructor syntax
This improves the error recovery for syntax like `new C.named();`. Previously, the parser would derail with messages like "A method declaration needs an explicit list of parameters" and a cascade of other errors.

With this change the qualified name is recognized as an attemp to write the name of the constructor, reporting that qualified names are not allowed in this case.

Part of https://github.com/dart-lang/sdk/issues/61699

Change-Id: Id1062aac3b90db8d2346013cf8a0bd9a541b9a3c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463020
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>