[messages] Fully remove class ParserErrorCode.
This class only remained in order to support version 3.1.3 of
`package:dart_style`, which was importing analyzer private sources and
referencing it. Versions of `package:dart_style` after 3.1.3 will not
need to refer to this class, thanks to
https://github.com/dart-lang/dart_style/commit/e8190bf2242654daee7ebf21fd6d8c8046989822.

The pubspec of version 3.1.3 of `package:dart_style` refers to
`package:analyzer` using a consraint of `>=8.2.0 <10.0.0`, so now that
`package:analyzer` has been bumped to version `10.0.0-dev`, this code
is no longer needed.

Change-Id: I6a6a696437a3b514cea4e33d231d97bef16d259c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/462340
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>