[analyzer] Make unignorable_ignore logic testable.
For some time, the analyzer logic for reporting the
`unignorable_ignore` warning has been disabled, with a comment
indicating that violations in the flutter codebase need to be cleaned
up before it can be re-enabled.

This change modifies the method by which this logic is disabled:
instead of commenting it out, it is disabled through the use of a
static boolean flag. This allows the logic to be tested by analyzer
unit tests.

I believe (based on trybot testing) that there are no remaining
violations in the flutter codebase that need to be cleaned up, and so
this logic can be re-enabled. However, just in case there's something
I'm missing, I will re-enable the logic in a follow-up CL that can be
more easily reverted.

Change-Id: I6a6a6964ce49b9aaa5fcd01ed0c86ee545c071c3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463402
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>