[analyzer] Enable unignorable_ignore diagnostic.
For some time, the analyzer logic for reporting the
`unignorable_ignore` warning has been disabled, with a comment
indicating that violations in the flutter codebase need to be cleaned
up before it can be re-enabled.

I believe (based on trybot testing) that there are no remaining
violations in the flutter codebase that need to be cleaned up, and so
this logic has been re-enabled. The flag that allows it to be disabled
is still in place, so that it can easily be disabled again if problems
crop up.

Once this change rolls into the Flutter codebase and I'm sufficiently
convinced that it hasn't led to failures, I'll remove the flag.

Change-Id: I6a6a6964e70e0e36247e9b02c43d4f819eaf5d87
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463504
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>