[linter] Test unnecessary ignore logic for unique names.
In https://dart-review.googlesource.com/c/sdk/+/463083, I removed some
logic from `ignore_validator.dart` that referred to diagnostic code
unique names, believing it to be unnecessary because `ignore` comments
couldn't use unique names. It turns out that the code *was*
unnecessary, but for an entirely different reason: because `ignore`
comments that use unique names are never reported by the
`unnecessary_ignore` lint due to logic elsewhere in the
`IgnoreValidator`*.

My confusion would have been less if the unit tests for the
`unnecessary_ignore` lint actually tested that `unnecessary_ignore`
_isn't_ reported when an `ignore` comment *is* necessary. So I'm
adding those tests in this CL.

(*The logic is the line `if
(!_validDiagnosticCodeNames.contains(name.toLowerCase())) continue;`
in
`IgnoreValidator._reportUnnecessaryOrRemovedOrDeprecatedIgnores`. This
code attempts to suppress `unnecessary_ignore` comments for
unrecognized codes, in order to allow for `ignore` comments that apply
to plug-ins. But it winds up suppressing `unnecessary_ignore` comments
for unique names as well. I will fix this in a follow-up CL.)

Change-Id: I6a6a69647e2ed35665ac41749ee0f56e7d0a766a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/463920
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>