[linter] Remove the pseudo-diagnostic code `removedLint`.
This code existed solely so that no-longer-supported lints (which
nonetheless remained in the codebase so they could be recognized by
the analyzer as "removed") would have a diagnostic code to point to.

Now that the preferred mechanism for marking a lint as removed has
been changed to a mechanism that doesn't require a pseudo-diagnostic
code (see https://dart-review.googlesource.com/c/sdk/+/463463),
`removedLint` can be deleted.

Change-Id: I6a6a696450e6583f858fd94ea315a34fbd4e3c54
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/461800
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>